requirements.txt
boto3==1.35.0
python-dateutil==2.9.0
pyyaml==6.0.1
python-dotenv==1.0.1
pytest==8.3.2
pytest-cov==5.0.0
moto==5.0.12
black==24.8.0
pylint==3.2.6